+++ package.json
@@ -7,6 +7,7 @@
     "type": "git",
     "url": "git@github.com:htmlacademy-univer-javascript-1/1199123-kekstagram-3.git"
   },
+  "type": "module",
   "bugs": {
     "url": "https://github.com/htmlacademy-univer-javascript-1/1199123-kekstagram-3/issues"
   },
